@@ -1,13 +1,13 @@
-const {
+import {
   LoginPrompt,
-}                = require('./Misc');
-const React      = require('react');
-const PropTypes  = require('prop-types');
-const classNames = require('classnames');
-const  Footer    = require('./Footer');
-const  ReactDOM  = require('react-dom');
-const  Sefaria   = require('./sefaria/sefaria');
-const  $         = require('./sefaria/sefariaJquery');
+} from './Misc';
+import React  from 'react';
+import PropTypes  from 'prop-types';
+import classNames  from 'classnames';
+import Footer  from './Footer';
+import ReactDOM  from 'react-dom';
+import Sefaria  from './sefaria/sefaria';
+import $  from './sefaria/sefariaJquery';
 import Component from 'react-class';
 
 
@@ -87,4 +87,4 @@ NotificationsPanel.propTypes = {
 };
 
 
-module.exports = NotificationsPanel;
+export default NotificationsPanel;
