@@ -1,4 +1,4 @@
-const {
+import {
   ReaderNavigationMenuCloseButton,
   ReaderNavigationMenuDisplaySettingsButton,
   CategoryAttribution,
@@ -6,14 +6,14 @@ const {
   LoadingMessage,
   TwoBox,
   LoginPrompt,
-}                = require('./Misc');
-const React      = require('react');
-const ReactDOM   = require('react-dom');
-const $          = require('./sefaria/sefariaJquery');
-const Sefaria    = require('./sefaria/sefaria');
-const classNames = require('classnames');
-const PropTypes  = require('prop-types');
-const sanitizeHtml = require('sanitize-html');
+} from './Misc';
+import React  from 'react';
+import ReactDOM  from 'react-dom';
+import $  from './sefaria/sefariaJquery';
+import Sefaria  from './sefaria/sefaria';
+import classNames  from 'classnames';
+import PropTypes  from 'prop-types';
+import sanitizeHtml  from 'sanitize-html';
 import Component from 'react-class';
 
 
@@ -317,4 +317,4 @@ SheetMetadata.propTypes = {
 
 
 
-module.exports = SheetMetadata;
+export default SheetMetadata;
