@@ -1,4 +1,4 @@
-const {
+import {
   SheetTagLink,
   SheetAccessIcon,
   CategoryColorLine,
@@ -6,15 +6,15 @@ const {
   LoadingMessage,
   TwoOrThreeBox,
   ProfilePic,
-}                = require('./Misc');
-const React      = require('react');
-const ReactDOM   = require('react-dom');
-const $          = require('./sefaria/sefariaJquery');
-const Sefaria    = require('./sefaria/sefaria');
-const classNames = require('classnames');
-const PropTypes  = require('prop-types');
-const GroupPage  = require('./GroupPage');
-const Footer     = require('./Footer');
+} from './Misc';
+import React  from 'react';
+import ReactDOM  from 'react-dom';
+import $  from './sefaria/sefariaJquery';
+import Sefaria  from './sefaria/sefaria';
+import classNames  from 'classnames';
+import PropTypes  from 'prop-types';
+import GroupPage  from './GroupPage';
+import Footer  from './Footer';
 import Component from 'react-class';
 
 
@@ -616,4 +616,4 @@ PrivateSheetListing.propTypes = {
 };
 
 
-module.exports = SheetsNav;
+export default SheetsNav;
