@@ -1,19 +1,19 @@
-const {
+import {
   CategoryColorLine,
   ReaderNavigationMenuMenuButton,
   ReaderNavigationMenuDisplaySettingsButton,
   LanguageToggleButton,
   LoadingMessage,
   Link,
-}                         = require('./Misc');
-const React               = require('react');
-const PropTypes           = require('prop-types');
-const ReactDOM            = require('react-dom');
-const classNames          = require('classnames');
-const Sefaria             = require('./sefaria/sefaria');
-const $                   = require('./sefaria/sefariaJquery');
-const TextRange           = require('./TextRange');
-const Footer              = require('./Footer');
+} from './Misc';
+import React  from 'react';
+import PropTypes  from 'prop-types';
+import ReactDOM  from 'react-dom';
+import classNames  from 'classnames';
+import Sefaria  from './sefaria/sefaria';
+import $  from './sefaria/sefariaJquery';
+import TextRange  from './TextRange';
+import Footer  from './Footer';
 import Component          from 'react-class';
 
 
@@ -175,4 +175,4 @@ TopicSource.propTypes = {
 }
 
 
-module.exports = TopicPage;
+export default TopicPage;
