@@ -149,4 +149,4 @@ class FilterNode {
   }
 }
 
-module.exports = FilterNode;
+export default FilterNode;
