@@ -105,4 +105,4 @@ class Track {
     }
 }
 
-module.exports = Track;
+export default Track;
